+++ src/controllers/middlewares/expenses.validation.js
@@ -1,12 +1,12 @@
 const stringIsFloat = require("../utils/validation/stringIsFloat");
 const stringIsInteger = require("../utils/validation/stringIsInteger");
-const getMonth = require("../../src/utils/getMonth");
+const getMonth = require("../../utils/getMonth");
 const isValidType = require("../utils/validation/validType");
 const isValidTransactionType = require("../utils/validation/isValidTransactionType");
-const expenseModel = require("../../models/expense.model");
-const incomeModel = require("../../models/income.model");
-const monthModel = require("../../models/month.model");
-const setUpModel = require("../../models/setup.model");
+const expenseModel = require("../../../models/expense.model");
+const incomeModel = require("../../../models/income.model");
+const monthModel = require("../../../models/month.model");
+const setUpModel = require("../../../models/setup.model");
 
 /**
  * Validates all the data for addExpense in expenses.controller
