+++ src/controllers/utils/validation/isValidTransactionType.js
@@ -1,4 +1,4 @@
-const transactionTypes = require('../../../src/utils/transactionTypes');
+const transactionTypes = require('../../../utils/transactionTypes');
 
 function isValidTransactionType(transaction) {
     return transaction === transactionTypes.EXPENSE || transaction === transactionTypes.INCOME;
